+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.0.1-alpha.322](https://github.com/fluxopt/lpspec/compare/v0.0.1-alpha.321...v0.0.1-alpha.322) (2026-09-12)
+
+
+### Features
+
+* **api:** an archive records how big the model was and how long it took ([#1598](https://github.com/fluxopt/lpspec/issues/1598)) ([80f6a6e](https://github.com/fluxopt/lpspec/commit/80f6a6efd11fffc50242626bd20db722a401a772))
+
+## [0.0.1-alpha.321](https://github.com/fluxopt/lpspec/compare/v0.0.1-alpha.320...v0.0.1-alpha.321) (2026-09-12)
+
+
+### Documentation
+
+* **bench:** the benchmark page is re-taken from the newest run, its claims about the numbers included ([#1587](https://github.com/fluxopt/lpspec/issues/1587)) ([0210774](https://github.com/fluxopt/lpspec/commit/02107748aba0fd82b1b22c298a9bbe187e8ec91d))
+
 ## [0.0.1-alpha.320](https://github.com/fluxopt/lpspec/compare/v0.0.1-alpha.319...v0.0.1-alpha.320) (2026-09-11)